requirements.txt
# Core
torch
torchvision
torchsummary
numpy 
scipy
matplotlib

# Image processing
Pillow
opencv-python
opencv-contrib-python
opencv-python-headless

# Training
wandb
tqdm 

# Utiltiies
pyyaml
click

# Testing
pytest